# [V0] Correct CUDA Graph capture for encoder-decoder models