--- a/vllm/config/__init__.py
+++ b/vllm/config/__init__.py
@@ -1166,15 +1166,15 @@ class ModelConfig:
 
     def _verify_cuda_graph(self) -> None:
         # The `max_seq_len_to_capture` was incorrectly
-        # based on the encoder's input length (448) 
-        # but not the decoder's larger input length (1500). 
+        # based on the encoder's input length (448)
+        # but not the decoder's larger input length (1500).
         # This change ensures the CUDA Graph captures the correct,
         # larger sequence length, allowing it to work as intended.
         effective_max_seq_len = self.max_model_len
         if self.is_encoder_decoder:
             effective_max_seq_len = max(
-                effective_max_seq_len, 
-              getattr(self.hf_config, "max_source_positions", 0))
+                effective_max_seq_len,
+                getattr(self.hf_config, "max_source_positions", 0))
         self.max_seq_len_to_capture = min(self.max_seq_len_to_capture,
                                           effective_max_seq_len)
         # CUDAGraph capture not supported for enc-dec models and mllama on ROCm