@Sugar-zsg 2025-08-11
> Have you considered the impact of this change on mllama? It would be great if you run benchmarks on both whisper and mllama

Thanks for the suggestion.

I have verified the impact on the Whisper-large-v3-turbo model using NVIDIA H20e GPUs. The change reduced the average inference time from approximately 320ms to 190ms for the same audio, which is a significant performance gain. This updated Whisper service has been running successfully in our production environment under high load for several weeks.

Regarding mllama, I haven't used it before, but I will start some basic performance and load tests to evaluate the impact.