@Sugar-zsg 2025-08-11
re: @DarkLight1337 

Thanks again for the suggestion and for your patience.

I have completed the benchmarks for **mllama** and can now share the full findings.

For the **meta-llama/Llama-3.2-11B-Vision-Instruct** model, tested on **NVIDIA H20e GPUs**, the performance remained stable:

* **Before the change:**
    * Token Throughput: 102.25 tokens/s
    * Average Latency: 2.504s/sample
* **After the change:**
    * Token Throughput: 102.09 tokens/s
    * Average Latency: 2.508s/sample

This stability is expected, as the `mllama` model did not have the same CUDA Graph configuration issue that the `whisper` model had. CUDA Graph was already correctly enabled for mllama, so my fix did not impact its performance.

As a reminder, this fix provided a significant performance boost for **Whisper-large-v3-turbo**, reducing average latency from ~320ms to 190ms. This change is stable and has been running in our production environment for weeks.

I believe this change is valuable for models like Whisper, where the encoder-decoder sequence length mismatch prevents CUDA Graph from being fully utilized.