@Sugar-zsg 2025-08-12
I've resolved the issue with the worker test and added some comments.